计算multip

+++ app/controllers/user_logos_controller.rb
@@ -9,6 +9,10 @@ class UserLogosController < ApplicationController
     user_logo = UserLogo.new(params[:user_logo])
     user_logo.user_id = session[:user_id]
     user_logo.save!
+    user = user_logo.user
+    unless user.multip
+      user.set_if_multip
+    end
     render :json => user_logo.output_hash.to_json
   end
   
@@ -43,7 +47,13 @@ class UserLogosController < ApplicationController
       render :json => {:error => "photo's owner #{user_logo.user_id} != session user #{session[:user_id]}"}.to_json
       return
     end
+    countp = user_logo.user.user_logos.count()
+    if countp<=1
+      render :json => {:error => "must have at least one photo"}.to_json
+      return
+    end
     if user_logo.destroy
+      user.update_attributes!({multip:false}) if countp<=2
       render :json => {:deleted => params[:id]}.to_json
     else
       render :json => {:error => "user_logo #{params[:id]} delete failed"}.to_json
